requirements.txt
absl-py==0.9.0
appdirs==1.4.4
boto3==1.14.11
botocore==1.17.11
cachetools==4.1.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cityscapesScripts==1.5.0
click==7.1.2
cycler==0.10.0
Cython==0.29.20
dataclasses==0.7
docutils==0.15.2
filelock==3.0.12
google-auth==1.18.0
google-auth-oauthlib==0.4.1
grpcio==1.30.0
idna==2.9
importlib-metadata==1.7.0
jmespath==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
ninja==1.10.0.post1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
packaging==20.4
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
sentencepiece==0.1.91
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorboardX==2.0
tokenizers==0.8.1rc1
torchvision==0.2.2
tqdm==4.46.1
transformers==3.0.2
urllib3==1.25.9
Werkzeug==1.0.1
yacs==0.1.7
zipp==3.1.0
